refactor: delete mvc-refactor folder + fix(test)
